[#3190] * Add simple unit test * Add Dockerfile.test and associated scripts to run tests inside docker container and copy the results out * Update Jenkinsfile to invoke ci test script * Refactor some duplicated code into analytics_helpers
There were some complications with the initial implementations of this
work, stemming from an attempt to run the tests inside a docker container,
inside Jenkins, and then using a bind mount to capture results so that
they could be processed further by Jenkins. Because the python container
runs as root by default, this created files in the Jenkins host owned
by root that Jenkins itself could not delete. The fix was to avoid bind
mounting, and instead copy the results out once the container was done
running, then delete the container (this is in run-ci-tests.sh)
