+++ .coveragerc
@@ -1,0 +1,13 @@
+[run]
+branch = True
+include =
+  scripts/*
+omit =
+  **/test_*
+  **/conftest*
+
+[html]
+directory = results/coverage
+
+[xml]
+output = results/coverage/pytest-coverage.xml
