+++ Dockerfile.test
@@ -1,0 +1,16 @@
+FROM python:3.6.8
+
+WORKDIR scorecard
+
+COPY scripts/requirements.txt     scripts/requirements.txt
+COPY scripts/dev-requirements.txt scripts/dev-requirements.txt
+
+RUN python3 -m pip install --trusted-host pypi.org --trusted-host files.pythonhosted.org --require-hashes \
+    --no-cache-dir --no-deps -r scripts/requirements.txt -r scripts/dev-requirements.txt
+
+COPY .coveragerc .
+COPY pytest.ini .
+COPY scripts/ scripts
+COPY test/ test
+
+CMD python -m pytest --cov --cov-report term --cov-report html --cov-report xml
