+++ pytest.ini
@@ -1,2 +1,8 @@
 [pytest]
-junit_family=xunit1
+junit_family=xunit2
+addopts =
+  --tb=short
+  # Show extra test summary info as specified by chars (f)ailed, (E)error, (s)skipped, (x)failed, (X)passed.
+  -rfEsxX
+  # Output test results to junit.xml for Jenkins to consume
+  --junitxml=results/unit/pytest-unit.xml
