+++ scripts/dev-requirements.in
@@ -5,6 +5,6 @@
 
 # Dev tooling
 flake8
-virtualenv
+freezegun
 pip-tools
 pytest-cov
