+++ scripts/requirements.txt
@@ -4,13 +4,13 @@
 #
 #    pip-compile --generate-hashes requirements.in
 #
-boto3==1.12.14 \
-    --hash=sha256:3b00b93d9548ebc0777f54cbd77c11e2721f7bb873ffb371cf4fa66d06750de7 \
-    --hash=sha256:3d3973502f96faaa534faec083ce4cd51a969f613d740be9e34cb7e42238f59c \
+boto3==1.12.17 \
+    --hash=sha256:13ac5526d80e1f1c9f7adcaeae7237f8bd1a5f268e918ca16e770eb865d3292b \
+    --hash=sha256:ef2d050e0c37e00dab6ff1135befe7ed4b9e1ec26c752b8ad01d043e184ea6d1 \
     # via credstash
-botocore==1.15.14 \
-    --hash=sha256:37e713b86bc833587e002610d0b2b5c5c88438617d68189d1b541e20726a43d4 \
-    --hash=sha256:4f889842519f54d007e285637c625a0645ffbe6aaa291b62ba510f3a25948f77 \
+botocore==1.15.17 \
+    --hash=sha256:17a70c5d8743237896907d00f3de0f51d31aeca863ab0da8d0d8c178608b7ce1 \
+    --hash=sha256:b09aed94cd7a01fbccbbf5746b69ca62ed37c0feaed42e0420f9a121ae7df755 \
     # via boto3, s3transfer
 cachetools==4.0.0 \
     --hash=sha256:9a52dd97a85f257f4e4127f15818e71a0c7899f121b34591fcc1173ea79a0198 \
